[Feature][Messaging] Thread deletion

+++ XClone/Core/Features/Messaging/Inbox/Service/InboxService.swift
@@ -16,7 +16,6 @@ protocol InboxServiceProtocol {
 
 class InboxService: InboxServiceProtocol {
     private var firestoreListener: ListenerRegistration?
-    private var threads = [Thread]()
     
     deinit {
         self.firestoreListener?.remove()
@@ -29,68 +28,46 @@ class InboxService: InboxServiceProtocol {
         await withThrowingTaskGroup(of: Void.self) { group in
             group.addTask { try await self.removeUserFromThread(thread, uid: uid) }
             group.addTask { try await self.deleteThreadIfNecessary(thread) }
+            group.addTask { try await self.updateThreadDeletionStructures(thread) }
+
         }
     }
     
     func fetchThreads() async throws -> [Thread] {
-        guard let currentUid = Auth.auth().currentUser?.uid else { return [] }
+        guard let threads = try await threadsQuery?.getDocuments(as: Thread.self) else { return [] }
         
-        let snapshot = try await FirestoreConstants
-            .ThreadsCollection
-            .whereField("uids", arrayContains: currentUid)
-            .order(by: "lastUpdated", descending: true)
-            .getDocuments()
-        
-        self.threads = snapshot.documents
-            .compactMap { try? $0.data(as: Thread.self) }
-            .sorted { thread1, thread2 in
-                let isFirstUnread = thread1.lastMessage?.status != .read
-                let isSecondUnread = thread2.lastMessage?.status != .read
-                
-                if isFirstUnread != isSecondUnread {
-                    return isFirstUnread // Prioritize unread messages
-                }
-                
-                return (thread1.lastMessage?.timestamp ?? .distantPast) >
-                       (thread2.lastMessage?.timestamp ?? .distantPast)
+        return threads.sorted { thread1, thread2 in
+            let isFirstUnread = thread1.lastMessage?.status != .read
+            let isSecondUnread = thread2.lastMessage?.status != .read
+            
+            if isFirstUnread != isSecondUnread {
+                return isFirstUnread // Prioritize unread messages
             }
-        
-        return threads
+            
+            return (thread1.lastMessage?.timestamp ?? .distantPast) >
+            (thread2.lastMessage?.timestamp ?? .distantPast)
+        }
     }
     
     func threadStream() -> AsyncStream<Thread> {
         return AsyncStream { continuation in
-            guard let currentUid = Auth.auth().currentUser?.uid else { return }
-            
             continuation.onTermination = { [weak self] _ in
                 self?.firestoreListener?.remove()
                 self?.firestoreListener = nil
                 continuation.finish()
             }
             
-            self.firestoreListener = FirestoreConstants
-                .ThreadsCollection
-                .whereField("uids", arrayContains: currentUid)
-                .order(by: "lastUpdated", descending: true)
-                .addSnapshotListener { snapshot, _ in
-                    guard let snapshot else { return }
-                    
-                    let threads = snapshot.documentChanges
-                        .filter { $0.type == .added || $0.type == .modified }
-                        .compactMap { try? $0.document.data(as: Thread.self) }
-                    
-                    if let thread = threads.first {
-                        continuation.yield(thread)
-                    }
+            self.firestoreListener = threadsQuery?.addSnapshotListener { snapshot, _ in
+                guard let snapshot else { return }
+                
+                let threads = snapshot.documentChanges
+                    .filter { $0.type == .added || $0.type == .modified }
+                    .compactMap { try? $0.document.data(as: Thread.self) }
+                
+                if let thread = threads.first {
+                    continuation.yield(thread)
                 }
-        }
-    }
-}
-
-private extension InboxService {
-    func deleteThreadIfNecessary(_ thread: Thread) async throws {
-        if thread.uids.count == 1 {
-            try await FirestoreConstants.ThreadsCollection.document(thread.id).delete()
+            }
         }
     }
 }
@@ -107,11 +84,20 @@ private extension InboxService {
     
     func removeUserFromThread(_ thread: Thread, uid: String) async throws {
         try await FirestoreConstants.ThreadsCollection.document(thread.id).updateData([
-           "uids": FieldValue.arrayRemove([uid])
+            "uids": FieldValue.arrayRemove([uid])
         ])
     }
     
-    func updateThreadDeletionStructures(_ thread: Thread, with chatPartnerID: String) async throws {
+    func deleteThreadIfNecessary(_ thread: Thread) async throws {
+        if thread.uids.count == 1 {
+            try await FirestoreConstants.ThreadsCollection.document(thread.id).delete()
+        }
+    }
+    
+    func updateThreadDeletionStructures(_ thread: Thread) async throws {
+        guard let currentUid = Auth.auth().currentUser?.uid else { return }
+        guard let chatPartnerID = thread.chatPartnerID(currentUserID: currentUid) else { return }
+        
         if thread.uids.isEmpty {
             try await FirestoreConstants.ThreadsCollection.document(thread.id).delete()
